+++ src/routes/_unregistered/shifts.view.tsx
@@ -1,16 +1,14 @@
-import { Flex, Icon, Text, VStack } from "@chakra-ui/react";
 import { createFileRoute } from "@tanstack/react-router";
-import { useMutation, useQuery } from "convex/react";
-import { useEffect, useRef, useState } from "react";
-import { LuTriangleAlert } from "react-icons/lu";
+import { useQuery } from "convex/react";
 import { api } from "@/convex/_generated/api";
 import type { Id } from "@/convex/_generated/dataModel";
 import { ExpiredView } from "@/src/components/features/StaffView/ExpiredView";
+import { RateLimitedView } from "@/src/components/features/StaffView/RateLimitedView";
 import { ShiftViewPage } from "@/src/components/features/StaffView/ShiftViewPage";
 import { StaffLayout } from "@/src/components/templates/StaffLayout";
 import { ErrorBoundary } from "@/src/components/ui/ErrorBoundary";
 import { FullPageSpinner } from "@/src/components/ui/FullPageSpinner";
-import { clearSession, getStoredSession, type SessionInfo, storeSession } from "@/src/utils/staffSession";
+import { useStaffSession } from "@/src/hooks/useStaffSession";
 
 export const Route = createFileRoute("/_unregistered/shifts/view")({
   validateSearch: (search: Record<string, unknown>) => ({
@@ -21,103 +19,43 @@ export const Route = createFileRoute("/_unregistered/shifts/view")({
 
 function ShiftViewRoute() {
   const { token } = Route.useSearch();
-  const verifyToken = useMutation(api.staffAuth.mutations.verifyToken);
+  const state = useStaffSession(token);
 
-  // LSにセッションがあればそちらを優先（トークンはワンタイムのため再検証しない）
-  const initialSession = getStoredSession();
-  const [session, setSession] = useState<SessionInfo | null>(initialSession);
-  const [expired, setExpired] = useState<{ recruitmentId: string | null } | null>(
-    !token && !initialSession ? { recruitmentId: null } : null,
-  );
-  const [rateLimited, setRateLimited] = useState(false);
-  const [verifying, setVerifying] = useState(false);
-  const verifyingRef = useRef(false);
-
-  useEffect(() => {
-    if (!token || session || verifyingRef.current) return;
-
-    verifyingRef.current = true;
-    setVerifying(true);
-
-    verifyToken({ token })
-      .then((result) => {
-        if (result.status === "ok") {
-          storeSession(result.recruitmentId, result.sessionToken);
-          setSession({ sessionToken: result.sessionToken, recruitmentId: result.recruitmentId });
-        } else if (result.status === "rate_limited") {
-          setRateLimited(true);
-        } else {
-          setExpired({ recruitmentId: result.recruitmentId });
-        }
-      })
-      .catch(() => {
-        setExpired({ recruitmentId: null });
-      })
-      .finally(() => {
-        verifyingRef.current = false;
-        setVerifying(false);
-      });
-  }, [token, session, verifyToken]);
-
-  if (rateLimited) {
+  if (state.status === "loading") return <FullPageSpinner />;
+  if (state.status === "rateLimited") return <RateLimitedView title="シフト閲覧" />;
+  if (state.status === "expired") {
     return (
       <StaffLayout shopName="シフト閲覧">
-        <Flex flex={1} align="center" justify="center" px={8}>
-          <VStack gap={4}>
-            <Icon boxSize={12} color="orange.500">
-              <LuTriangleAlert />
-            </Icon>
-            <Text fontSize="lg" fontWeight="semibold" textAlign="center">
-              アクセス制限中
-            </Text>
-            <Text fontSize="sm" color="fg.muted" textAlign="center">
-              しばらく時間を置いてから{"\n"}再度アクセスしてください
-            </Text>
-          </VStack>
-        </Flex>
+        <ExpiredView recruitmentId={state.recruitmentId} />
       </StaffLayout>
     );
-  }
-
-  if (expired) {
-    return (
-      <StaffLayout shopName="シフト閲覧">
-        <ExpiredView recruitmentId={expired.recruitmentId} />
-      </StaffLayout>
-    );
-  }
-
-  if (!session || verifying) {
-    return <FullPageSpinner />;
   }
 
   return (
     <ErrorBoundary
       fallback={
         <StaffLayout shopName="シフト閲覧">
-          <ExpiredView recruitmentId={session.recruitmentId} />
+          <ExpiredView recruitmentId={state.session.recruitmentId} />
         </StaffLayout>
       }
       onError={(error) => {
         if (error.message?.includes("ArgumentValidationError")) {
-          clearSession(session.recruitmentId);
+          state.clearSession();
         }
       }}
     >
-      <ShiftViewContent session={session} />
+      <ShiftViewContent session={state.session} />
     </ErrorBoundary>
   );
 }
 
-function ShiftViewContent({ session }: { session: SessionInfo }) {
+function ShiftViewContent({ session }: { session: { sessionToken: string; recruitmentId: string } }) {
   const data = useQuery(api.staffAuth.queries.getShiftViewData, {
     sessionToken: session.sessionToken,
     recruitmentId: session.recruitmentId as Id<"recruitments">,
   });
 
-  if (data === undefined) {
-    return <FullPageSpinner />;
-  }
+  if (data === undefined) return <FullPageSpinner />;
 
   if (data === null) {
     return (
